refactor: getUserData 함수에서 getUserLoginData로 변경하여 로그인한 사용자 데이터 가져오기
