+++ app/manager/apply/page.tsx
@@ -1,10 +1,10 @@
-import { getUserData } from '@/src/shared/api/getUserData';
+import { getUserLoginData } from '@/src/shared/api/getUserData';
 import MyGoogleCaptcha from '@/src/shared/ui/GoogleRecaptcha';
 
 import { ManagerApplyPage } from './ui/ManagerApplyPage';
 
 export default async function Page() {
-  const userData = await getUserData();
+  const userData = await getUserLoginData();
   return (
     <MyGoogleCaptcha>
       <ManagerApplyPage userData={userData.userData} />
